Move shared rule-expression functions to neutral manager.expression package
The CLAIM / CLAIM_IS_VERIFIED / CLIENT EvalEx functions are generic building
blocks used by both the scope-granting rules (manager.rule) and the act-as rules
(manager.actas). Move them to a concern-neutral com.sympauthy.business.manager.expression
package so manager.actas no longer depends on the scope-granting package. Each
concern keeps its own executor, function registration, validation and exception.
